chore(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics
The variable `port` in the `server.ts` file was changed to `PORT` to improve the semantics of the code. This change was made to follow conventional naming conventions and make the code more readable.

feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port

In the `server.ts` file, support for the `process.env.PORT` environment variable was added. This allows the application to be run on a configurable port specified by the environment variable. This change was made to provide flexibility in deploying the application and to make it easier to configure the port without modifying the code directly.

fix(Carousel.svelte): adjust height and position of carousel elements

In the `Carousel.svelte` file, the height and position of the carousel elements were adjusted. The height of the carousel container was changed from `h-80` to `h-[26rem]` to provide a better visual experience. Additionally, the position of the text elements within the carousel was adjusted from `top-24` to `bottom-24` to align them at the bottom of the carousel. These changes were made to improve the layout and aesthetics of the carousel component.

fix(Navbar.svelte): add class to prevent horizontal overflow

In the `Navbar.svelte` file, a class `overflow-x-hidden` was added to the `svelte:window` element. This class prevents horizontal overflow on the page. This change was made to ensure that the page content does not extend beyond the viewport horizontally.

fix(Navbar.svelte): add background blur effect to nav-top class

In the `Navbar.svelte` file, a background blur effect was added to the `nav-top` class. This gives a visually appealing effect to the navbar when it is scrolled. This change was made to enhance the visual experience of the navbar.

fix(SearchBar.svelte): reset input value after search

In the `SearchBar.svelte` file, the input value is now reset after performing a search. This ensures that the input field is cleared after a search is executed. This change was made to provide a better user experience and prevent the input field from retaining the previous search query.

fix(+page.svelte): comment out under construction alert and adjust container width

In the `+page.svelte` file, the under construction alert was commented out and